Have app connected, worked on polishing listing-details page.
