+++ Dev/Front-End/app-draft/.expo/devices.json
@@ -1,0 +1,3 @@
+{
+  "devices": []
+}
