+++ Dev/Front-End/app-draft/.expo/types/router.d.ts
@@ -1,0 +1,14 @@
+/* eslint-disable */
+import * as Router from 'expo-router';
+
+export * from 'expo-router';
+
+declare module 'expo-router' {
+  export namespace ExpoRouter {
+    export interface __routes<T extends string | object = string> {
+      hrefInputParams: { pathname: Router.RelativePathString, params?: Router.UnknownInputParams } | { pathname: Router.ExternalPathString, params?: Router.UnknownInputParams } | { pathname: `/add-listing-details`; params?: Router.UnknownInputParams; } | { pathname: `/`; params?: Router.UnknownInputParams; } | { pathname: `/login`; params?: Router.UnknownInputParams; } | { pathname: `/register`; params?: Router.UnknownInputParams; } | { pathname: `/_sitemap`; params?: Router.UnknownInputParams; } | { pathname: `${'/(tabs)'}/add-listing` | `/add-listing`; params?: Router.UnknownInputParams; } | { pathname: `${'/(tabs)'}` | `/`; params?: Router.UnknownInputParams; } | { pathname: `${'/(tabs)'}/userprofile/changepassword` | `/userprofile/changepassword`; params?: Router.UnknownInputParams; } | { pathname: `${'/(tabs)'}/userprofile` | `/userprofile`; params?: Router.UnknownInputParams; } | { pathname: `${'/(tabs)'}/userprofile/mylistings` | `/userprofile/mylistings`; params?: Router.UnknownInputParams; } | { pathname: `${'/(tabs)'}/userprofile/savedposts` | `/userprofile/savedposts`; params?: Router.UnknownInputParams; } | { pathname: `/context/SavedPostsContext`; params?: Router.UnknownInputParams; } | { pathname: `/+not-found`, params: Router.UnknownInputParams & {  } } | { pathname: `/sale/[id]`, params: Router.UnknownInputParams & { id: string | number; } };
+      hrefOutputParams: { pathname: Router.RelativePathString, params?: Router.UnknownOutputParams } | { pathname: Router.ExternalPathString, params?: Router.UnknownOutputParams } | { pathname: `/add-listing-details`; params?: Router.UnknownOutputParams; } | { pathname: `/`; params?: Router.UnknownOutputParams; } | { pathname: `/login`; params?: Router.UnknownOutputParams; } | { pathname: `/register`; params?: Router.UnknownOutputParams; } | { pathname: `/_sitemap`; params?: Router.UnknownOutputParams; } | { pathname: `${'/(tabs)'}/add-listing` | `/add-listing`; params?: Router.UnknownOutputParams; } | { pathname: `${'/(tabs)'}` | `/`; params?: Router.UnknownOutputParams; } | { pathname: `${'/(tabs)'}/userprofile/changepassword` | `/userprofile/changepassword`; params?: Router.UnknownOutputParams; } | { pathname: `${'/(tabs)'}/userprofile` | `/userprofile`; params?: Router.UnknownOutputParams; } | { pathname: `${'/(tabs)'}/userprofile/mylistings` | `/userprofile/mylistings`; params?: Router.UnknownOutputParams; } | { pathname: `${'/(tabs)'}/userprofile/savedposts` | `/userprofile/savedposts`; params?: Router.UnknownOutputParams; } | { pathname: `/context/SavedPostsContext`; params?: Router.UnknownOutputParams; } | { pathname: `/+not-found`, params: Router.UnknownOutputParams & {  } } | { pathname: `/sale/[id]`, params: Router.UnknownOutputParams & { id: string; } };
+      href: Router.RelativePathString | Router.ExternalPathString | `/add-listing-details${`?${string}` | `#${string}` | ''}` | `/${`?${string}` | `#${string}` | ''}` | `/login${`?${string}` | `#${string}` | ''}` | `/register${`?${string}` | `#${string}` | ''}` | `/_sitemap${`?${string}` | `#${string}` | ''}` | `${'/(tabs)'}/add-listing${`?${string}` | `#${string}` | ''}` | `/add-listing${`?${string}` | `#${string}` | ''}` | `${'/(tabs)'}${`?${string}` | `#${string}` | ''}` | `/${`?${string}` | `#${string}` | ''}` | `${'/(tabs)'}/userprofile/changepassword${`?${string}` | `#${string}` | ''}` | `/userprofile/changepassword${`?${string}` | `#${string}` | ''}` | `${'/(tabs)'}/userprofile${`?${string}` | `#${string}` | ''}` | `/userprofile${`?${string}` | `#${string}` | ''}` | `${'/(tabs)'}/userprofile/mylistings${`?${string}` | `#${string}` | ''}` | `/userprofile/mylistings${`?${string}` | `#${string}` | ''}` | `${'/(tabs)'}/userprofile/savedposts${`?${string}` | `#${string}` | ''}` | `/userprofile/savedposts${`?${string}` | `#${string}` | ''}` | `/context/SavedPostsContext${`?${string}` | `#${string}` | ''}` | { pathname: Router.RelativePathString, params?: Router.UnknownInputParams } | { pathname: Router.ExternalPathString, params?: Router.UnknownInputParams } | { pathname: `/add-listing-details`; params?: Router.UnknownInputParams; } | { pathname: `/`; params?: Router.UnknownInputParams; } | { pathname: `/login`; params?: Router.UnknownInputParams; } | { pathname: `/register`; params?: Router.UnknownInputParams; } | { pathname: `/_sitemap`; params?: Router.UnknownInputParams; } | { pathname: `${'/(tabs)'}/add-listing` | `/add-listing`; params?: Router.UnknownInputParams; } | { pathname: `${'/(tabs)'}` | `/`; params?: Router.UnknownInputParams; } | { pathname: `${'/(tabs)'}/userprofile/changepassword` | `/userprofile/changepassword`; params?: Router.UnknownInputParams; } | { pathname: `${'/(tabs)'}/userprofile` | `/userprofile`; params?: Router.UnknownInputParams; } | { pathname: `${'/(tabs)'}/userprofile/mylistings` | `/userprofile/mylistings`; params?: Router.UnknownInputParams; } | { pathname: `${'/(tabs)'}/userprofile/savedposts` | `/userprofile/savedposts`; params?: Router.UnknownInputParams; } | { pathname: `/context/SavedPostsContext`; params?: Router.UnknownInputParams; } | `/+not-found` | `/sale/${Router.SingleRoutePart<T>}` | { pathname: `/+not-found`, params: Router.UnknownInputParams & {  } } | { pathname: `/sale/[id]`, params: Router.UnknownInputParams & { id: string | number; } };
+    }
+  }
+}
